Fix timer bug (and end-of-game resource leak), puzzle-end-notification delay
cancelAnimationFrame (sometimes?) doesn't cancel a callback that has already
been scheduled in the current repaint. That was allowing onAnimationFrame
to restart a deactivated timer.

Change the logic to avoid keeping track of the current timerId, and don't bother
cancelling the pending timer in deactivateTimer; just ignore it in the callback.

Fixes timing of puzzle-end-notification in puzzles with both move and flash
animations. Likely also fixes the Safari forced reload after game end.